+++ bundles/core/src/test/java/com/adobe/cq/wcm/core/components/ContentFragmentModelsCommonsTest.java
@@ -19,12 +19,12 @@ import org.junit.Test;
 
 import com.adobe.cq.wcm.core.components.testing.AbstractModelTest;
 
-public class ModelsCommonsTest extends AbstractModelTest{
+public class ContentFragmentModelsCommonsTest extends AbstractModelTest {
 
     @Test
     public void testDefaultBehaviour() throws Exception {
         testDefaultBehaviour(new String[] {
-                "com.adobe.cq.wcm.core.components.extension.contentfragment.models"
+                "com.adobe.cq.wcm.core.components.models.v1.contentfragment"
         });
     }
