+++ bundles/core/src/test/resources/contentfragment/test-content-conf.json
