+++ bundles/core/src/test/resources/contentfragment/test-content-dam-collections.json
