+++ bundles/core/src/test/resources/contentfragment/test-content-dam-contentfragments.json
