+++ bundles/core/src/test/resources/contentfragment/test-content-datasources.json
