+++ bundles/core/src/test/resources/contentfragment/test-content-renderconditions.json
