+++ bundles/core/src/test/resources/contentfragment/test-content.json
