+++ extension/contentfragment/content/src/content/jcr_root/apps/core/wcm/extension/components/contentfragment/v1/contentfragment/rawcontent.html
@@ -13,4 +13,4 @@
   ~ See the License for the specific language governing permissions and
   ~ limitations under the License.
   ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-*/--><sly data-sly-use.fragment="com.adobe.cq.wcm.core.components.extension.contentfragment.models.ContentFragment">${fragment.elements[0].html @ context="unsafe"}</sly>
+*/--><sly data-sly-use.fragment="com.adobe.cq.wcm.core.components.models.contentfragment.ContentFragment">${fragment.elements[0].html @ context="unsafe"}</sly>
